requirements.txt
python-dotenv
httpx
gspread
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
apscheduler
pytz
openai
python-telegram-bot
